backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4
boto3==1.29.7
Pillow==10.1.0
django-cors-headers==4.3.1
python-decouple==3.8
dj-database-url==2.1.0
django-storages==1.14.2
